xivo_confd/helpers/converter.py
# -*- coding: utf-8 -*-

# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>

from werkzeug.routing import BaseConverter, ValidationError


class FilenameConverter(BaseConverter):

    def to_python(self, value):
        if not value or len(value) > 194:
            raise ValidationError()
        if u'/' in value or value.startswith(u'.'):
            raise ValidationError()
        return value
